API V3 DQ Rules - move Serializers
